arm64: dts: juno: add dtsi for HDMI daughterboard
The V2C-HDMI board is a daughterboard which plugs into the LogicTile on
a Juno development board, providing two HDMI outputs for use by devices
instantiated in the FPGA.

Add a device-tree for this daughterboard. It should be included as a
child of the i2c bus it is connected to.

Change-Id: Ic781abbfa7d579e16fe65c90ec0200e7ce00e728
Signed-off-by: Brian Starkey <brian.starkey@arm.com>