staging: android: ion: read coherent DMA memory from device tree
This change introduces for ion_dummy driver the ability to declare specific
DMA coherent memory in the device tree, using compatible = "ion,dma_heap";

If no node is found, the standard ARM DMA ops will be used.

If found, it'll declare the coherent memory as DMA_MEMORY_EXCLUSIVE, which means
that in case of memory shortage, the allocation will fail instead of falling
back to ARM DMA ops.

This work has mainly been based on the patch [1]. However, in order to
have a less intrusive change, the dummy ION driver is modified instead of
bringing a complete new driver which is in the end a copy of ion_dummy.
Moreover, the lack of support for device-tree in ION makes the creation of a
decent ION driver for Juno tricky.

[1] http://mpg-bsp.cambridge.arm.com:8080/#/c/3515

Change-Id: I461bf4d6e605af4cc3a45595f8233b1013985954